Refactor dataset page by componentizing renaming and captioning sections
Features:
- Created RenamingSection component with complete file renaming functionality
- Created CaptioningSection component with comprehensive JoyCaption integration
- Moved all renaming logic, state management, and UI to RenamingSection
- Moved all captioning logic, saved settings, and advanced options to CaptioningSection
- Cleaned up main dataset page by removing 400+ lines of embedded code
- Maintained all existing functionality while improving code organization
- Enhanced maintainability and reusability of both sections

Technical improvements:
- Proper component props interfaces for type safety
- Self-contained state management within each component
- Callback patterns for parent-child communication
- Removed unused imports and cleaned up code structure
- Better separation of concerns and single responsibility principle
